3e done, batch messages to send in new rebroadcast message type.
To reduce the average messages per op below our cluster size we have to
send more messages than one in a message.  So we accumulate messages for
some delay (500ms currently) and send them out in a single message to
other nodes.  Created a new message type for this so as not to overload
actual maelstrom types.

Since this is an all in one setup, the goal was to take our accumulate
delay to the absolute limit to accomplish both 3d and 3e, 500ms seems to
be that limit with our messages per op coming in just below 20 while
latencies coming in under 400 median 600 max

:servers {:send-count 35300,
         :recv-count 35300,
         :msg-count 35300,
         :msgs-per-op 19.556787}
:stable-latencies {0 0,
                   0.5 321,
                   0.95 487,
                   0.99 497,
                   1 581}